requirements.txt
# crawl and parse
feedparser
requests
bs4
lxml
json_repair

# storage related
python-dotenv
azure-storage-blob>=12.9.0
azure-storage-queue
azure-cosmos

# openai tools
openai
tiktoken

# embeddings
sentence-transformers

# specific tools
wikipedia

# website
django

# debugging
ipdb
